+++ resources/views/helper.php
@@ -71,6 +71,16 @@ namespace <?= $namespace == '__root' ? '' : trim($namespace, '\\') ?> {
 
 <?php endforeach; ?>
 
+<?php foreach($real_time_facades as $name): ?>
+<?php $nested = explode('\\', str_replace('\\'.class_basename($name), '', $name)); ?>
+namespace <?php echo implode('\\', $nested); ?> {
+    /**
+     * @mixin <?= str_replace('Facades', '', $name) ?>
+     */
+    class <?= class_basename($name) ?> extends <?= str_replace('Facades', '', $name) ?> {}
+}
+<?php endforeach; ?>
+
 <?php if ($helpers) : ?>
 namespace {
     <?= $helpers ?>